refactor: change name componet writercard
